requirements.txt
playwright==1.52.0
twilio==9.6.1
fastapi
uvicorn
python-dotenv==1.1.0
python-multipart==0.0.20